Use new test syntax for web-runtime pages
