+++ packages/web-test-helpers/src/mocks/store/defaultStoreMockOptions.ts
@@ -5,6 +5,8 @@ export const defaultStoreMockOptions = {
   getters: {
     newFileHandlers: jest.fn(() => []),
     activeNotifications: jest.fn(() => []),
+    getNavItemsByExtension: jest.fn(),
+    apps: jest.fn(() => ({})),
     quota: jest.fn(() => ({})),
     capabilities: jest.fn().mockImplementation(() => ({})),
     user: jest.fn().mockImplementation(() => ({})),
